copilot refactoring test. no functional request/response behavior was changed. new class/method structure created from old code
