+++ include/ServerEngine.hpp
@@ -1,28 +1,48 @@
-# pragma once
+#pragma once
 
-#include <poll.h>
-#include "ServerManager.hpp"
-#include "Client.hpp"
+# include <map>
+# include <poll.h>
+# include <ctime>
+# include <signal.h>
+# include <sys/wait.h>
 
-#define MAX_FDS 100
+# include "ServerManager.hpp"
+# include "Client.hpp"
+# include "CgiHandler.hpp"
+# include "HttpParser.hpp"
+# include "HttpException.hpp"
+# include "helperUtils.hpp"
 
-class ServerEngine {
+# define MAX_FDS 100
+# define POLL_TIMEOUT_MS 1000
+# define CLIENT_TIMEOUT 30
+# define CGI_TIMEOUT 10
+
+class ServerEngine
+{
     private:
-        
-        const ServerManager& _manager; 
+        const ServerManager& _manager;
+        std::map<int, Client> _clients;             // clientFd -> Client
+        std::map<int, CgiProcess> _cgiProcesses;    // cgi responseFd -> CgiProcess
+        std::map<int, int> _fdRegistry;             // cgi responseFd -> clientFd
+        std::map<int, const ServerConfig*> _masterSocketRegistry;
+        struct pollfd _fds[MAX_FDS];
+        HttpParser _httpParser;
 
-       //runtime registries
-        std::map<int, Client> _clients;
-        struct pollfd         _fds[MAX_FDS];
-        //void acceptNewClient(int master_fd);
-        //void readFromClient(int client_fd);
+        void printFdRegistry() const;
+        bool isClientWaitingForCgi(int clientFd) const;
+        void removeFdFromPoll(int fd);
+        void checkClientTimeouts(time_t now);
+        void checkCgiTimeouts(time_t now);
+
+        void acceptNewClient(int triggered_fd);
+        void handleCgiFd(int i, int triggered_fd);
+        void handleClientFd(int i, int currentFd);
 
     public:
+        ServerEngine(const ServerManager& manager,
+                     struct pollfd fds[],
+                     std::map<int, const ServerConfig*>& masterSocketRegistry);
 
-        // (constructor takes the manager as a parameter)
-        // intialize the poll loop to -1 heere
-        ServerEngine(const ServerManager& manager);
-
-        // (this is the function that starts the infinite while(true) loop)
-        void run(); 
+        void run();
 };
